requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
python-dotenv
requests
pydantic
huggingface_hub
ollama